requirements.txt
pefile
python-magic
yara-python
virustotal-api
gitpython
